# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.ProductInput:
    properties:
      brand:
        example: Royal Canin
        type: string
      category_id:
        example: 11111111-1111-1111-1111-111111111111
        type: string
      description:
        example: Premium dog food for medium breeds
        type: string
      image_url:
        example: https://example.com/image.jpg
        type: string
      name:
        example: Royal Canin Medium Adult
        type: string
      price:
        example: 1599
        type: number
      stock:
        example: 45
        type: integer
      weight:
        example: 3kg
        type: string
    required:
    - category_id
    - name
    - price
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@petfood.com
    name: API Support
  description: This is a Pet Food E-commerce API server with authentication, products,
    cart, and order management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Pet Food E-commerce API
  version: "1.0"
paths:
...
  /admin/products:
    post:
      consumes:
      - application/json
      description: Create a new product in the catalog
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/controllers.ProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product (Admin only)
      tags:
      - Admin - Products
  /admin/products/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a product from the catalog
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product (Admin only)
      tags:
      - Admin - Products
    put:
      consumes:
      - application/json
      description: Update an existing product's information
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/controllers.ProductInput'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a product (Admin only)
      tags:
      - Admin - Products
...
  /products:
    get:
      consumes:
      - application/json
      description: Get a paginated list of products with optional filtering by category
        and search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 12
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Filter by category ID
        in: query
        name: category_id
        type: string
      - description: Search by product name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of products with pagination info
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all products
      tags:
      - Products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
      summary: Get a product by ID
      tags:
      - Products
  /products/category/{categoryId}:
    get:
      consumes:
      - application/json
      description: Get all products belonging to a specific category
      parameters:
      - description: Category ID
        in: path
        name: categoryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of products in the category
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get products by category
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"